[CHORE] keywordTextField 컴포넌트화 진행중 (#212)
